Project/MATNet-ADA-Version/requirements.txt
easydict==1.9
lmdb==0.94
torchvision==0.2.2
opencv_contrib_python==4.0.0.21
scipy==1.2.1
# scipy
six==1.12.0
torch==1.0.1.post2
pydensecrf==1.0rc2
# tqdm==4.31.1
tqdm
scikit_image==0.15.0
joblib==0.13.2
numpy==1.16.2
matplotlib==3.0.3
# Pillow==8.1.1
Pillow
PyYAML==5.1.2
# skimage==0.0
scikit-image
